updated with changes

+++ database/migrations/2024_12_02_140724_create_images_table.php
@@ -10,7 +10,7 @@ return new class extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->id();
-            $table->integer('stock_id');
+            $table->unsignedBigInteger('stock_id');
             $table->string('image_path');
             $table->timestamps();
